rename `Comparam` to `ComparamInstance` and `*ComparamSpec` to `*Comparam`
ODX has a COMPARAM-SPEC element which is a container like DIAG-LAYER,
so naming the objects which define the fields `ComparamSpec` has too
much potential for confusion (even though IMO it is better than what
the ODX spec defines).

This was proposed by [at]kayoub5. Thanks!
