+++ odxtools/templates/macros/printComparamRef.xml.jinja2
@@ -2,7 +2,7 @@
  #
  # SPDX-License-Identifier: MIT
 -#}
-{%- import('macros/printComparamSpec.xml.jinja2') as pcp %}
+{%- import('macros/printComparam.xml.jinja2') as pcp %}
 {#- -#}
 
 {%- macro printComparamRef(cp) %}
