+++ tests/test_decoding.py
@@ -2,7 +2,7 @@
 import unittest
 from typing import cast
 
-from odxtools.comparam import Comparam
+from odxtools.comparaminstance import ComparamInstance
 from odxtools.compumethods.identicalcompumethod import IdenticalCompuMethod
 from odxtools.compumethods.limit import IntervalType, Limit
 from odxtools.compumethods.linearcompumethod import LinearCompuMethod
@@ -95,7 +95,7 @@ class TestIdentifyingService(unittest.TestCase):
             admin_data=None,
             semantic=None,
             audience=None,
-            comparam_refs=NamedItemList[Comparam](),
+            comparam_refs=NamedItemList[ComparamInstance](),
             is_cyclic_raw=None,
             is_multiple_raw=None,
             addressing_raw=None,
@@ -170,7 +170,7 @@ class TestIdentifyingService(unittest.TestCase):
             admin_data=None,
             semantic=None,
             audience=None,
-            comparam_refs=NamedItemList[Comparam](),
+            comparam_refs=NamedItemList[ComparamInstance](),
             is_cyclic_raw=None,
             is_multiple_raw=None,
             addressing_raw=None,
@@ -292,7 +292,7 @@ class TestDecoding(unittest.TestCase):
             is_executable_raw=None,
             is_final_raw=None,
             semantic=None,
-            comparam_refs=NamedItemList[Comparam](),
+            comparam_refs=NamedItemList[ComparamInstance](),
             is_cyclic_raw=None,
             is_multiple_raw=None,
             addressing_raw=None,
@@ -434,7 +434,7 @@ class TestDecoding(unittest.TestCase):
             is_final_raw=None,
             admin_data=None,
             semantic=None,
-            comparam_refs=NamedItemList[Comparam](),
+            comparam_refs=NamedItemList[ComparamInstance](),
             is_cyclic_raw=None,
             is_multiple_raw=None,
             addressing_raw=None,
@@ -618,7 +618,7 @@ class TestDecoding(unittest.TestCase):
             is_final_raw=None,
             admin_data=None,
             semantic=None,
-            comparam_refs=NamedItemList[Comparam](),
+            comparam_refs=NamedItemList[ComparamInstance](),
             is_cyclic_raw=None,
             is_multiple_raw=None,
             addressing_raw=None,
@@ -820,7 +820,7 @@ class TestDecoding(unittest.TestCase):
             is_final_raw=None,
             admin_data=None,
             semantic=None,
-            comparam_refs=NamedItemList[Comparam](),
+            comparam_refs=NamedItemList[ComparamInstance](),
             is_cyclic_raw=None,
             is_multiple_raw=None,
             addressing_raw=None,
@@ -977,7 +977,7 @@ class TestDecoding(unittest.TestCase):
             is_executable_raw=None,
             is_final_raw=None,
             semantic=None,
-            comparam_refs=NamedItemList[Comparam](),
+            comparam_refs=NamedItemList[ComparamInstance](),
             is_cyclic_raw=None,
             is_multiple_raw=None,
             addressing_raw=None,
@@ -1171,7 +1171,7 @@ class TestDecoding(unittest.TestCase):
             is_executable_raw=None,
             is_final_raw=None,
             semantic=None,
-            comparam_refs=NamedItemList[Comparam](),
+            comparam_refs=NamedItemList[ComparamInstance](),
             is_cyclic_raw=None,
             is_multiple_raw=None,
             addressing_raw=None,
